Move helpers related to functab and buildid to a new reflectext package
