+++ internal/reflectext/buildid.go
@@ -1,0 +1,8 @@
+package reflectext
+
+var buildID string
+
+// BuildID is the build identifier for the binary.
+func BuildID() string {
+	return buildID
+}
