+++ internal/reflectext/obj_darwin.go
@@ -1,4 +1,4 @@
-package types
+package reflectext
 
 import (
 	"bytes"
