+++ internal/reflectext/obj_linux.go
@@ -1,4 +1,4 @@
-package types
+package reflectext
 
 import (
 	"bytes"
