--- types/buildid.go
@@ -1,4 +1,0 @@
-package types
-
-// buildID is the build identifier for the binary.
-var buildID string
